refactor: split "meta" into `Meta`
