+++ src/awkward/_meta/unionmeta.py
@@ -1,0 +1,95 @@
+# BSD 3-Clause License; see https://github.com/scikit-hep/awkward/blob/main/LICENSE
+
+from __future__ import annotations
+
+from collections import Counter
+from collections.abc import Sequence
+
+from awkward._meta.meta import Meta
+from awkward._typing import JSONSerializable
+
+
+class UnionMeta(Meta):
+    is_union = True
+
+    _contents: Sequence[Meta]
+
+    def purelist_parameters(self, *keys: str) -> JSONSerializable:
+        if self._parameters is not None:
+            for key in keys:
+                if key in self._parameters:
+                    return self._parameters[key]
+
+        for key in keys:
+            out = self._contents[0].purelist_parameter(key)
+            for content in self._contents[1:]:
+                tmp = content.purelist_parameter(key)
+                if out != tmp:
+                    return None
+            return out
+
+        return None
+
+    @property
+    def purelist_isregular(self) -> bool:
+        for content in self._contents:
+            if not content.purelist_isregular:
+                return False
+        return True
+
+    @property
+    def purelist_depth(self) -> int:
+        out = None
+        for content in self._contents:
+            if out is None:
+                out = content.purelist_depth
+            elif out != content.purelist_depth:
+                return -1
+        return out
+
+    @property
+    def is_identity_like(self) -> bool:
+        return False
+
+    @property
+    def minmax_depth(self) -> tuple[int, int]:
+        if len(self._contents) == 0:
+            return (0, 0)
+        mins, maxs = [], []
+        for content in self._contents:
+            mindepth, maxdepth = content.minmax_depth
+            mins.append(mindepth)
+            maxs.append(maxdepth)
+        return (min(mins), max(maxs))
+
+    @property
+    def branch_depth(self) -> tuple[bool, int]:
+        if len(self._contents) == 0:
+            return (False, 1)
+        anybranch = False
+        mindepth = None
+        for content in self._contents:
+            branch, depth = content.branch_depth
+            if mindepth is None:
+                mindepth = depth
+            if branch or mindepth != depth:
+                anybranch = True
+            if mindepth > depth:
+                mindepth = depth
+        return (anybranch, mindepth)
+
+    @property
+    def fields(self) -> list[str]:
+        field_counts = Counter([f for c in self._contents for f in c.fields])
+        return [f for f, n in field_counts.items() if n == len(self._contents)]
+
+    @property
+    def is_tuple(self) -> bool:
+        return all(x.is_tuple for x in self._contents) and (len(self._contents) > 0)
+
+    @property
+    def dimension_optiontype(self) -> bool:
+        for content in self._contents:
+            if content.dimension_optiontype:
+                return True
+        return False
