+++ src/awkward/contents/indexedoptionarray.py
@@ -8,6 +8,7 @@ from collections.abc import Mapping, MutableMapping, Sequence
 import awkward as ak
 from awkward._backends.backend import Backend
 from awkward._layout import maybe_posaxis
+from awkward._meta.indexedoptionmeta import IndexedOptionMeta
 from awkward._nplikes.array_like import ArrayLike
 from awkward._nplikes.numpy import Numpy
 from awkward._nplikes.numpy_like import IndexType, NumpyMetadata
@@ -50,7 +51,7 @@ numpy = Numpy.instance()
 
 
 @final
-class IndexedOptionArray(Content):
+class IndexedOptionArray(IndexedOptionMeta, Content):
     """
     IndexedOptionArray is an #ak.contents.IndexedArray for which
     negative values in the index are interpreted as missing. It represents
@@ -98,8 +99,7 @@ class IndexedOptionArray(Content):
                     raise AssertionError(where)
     """
 
-    is_option = True
-    is_indexed = True
+    _content: Content
 
     def __init__(self, index, content, *, parameters=None):
         if not (
