+++ src/awkward/contents/listarray.py
@@ -8,6 +8,7 @@ from collections.abc import Mapping, MutableMapping, Sequence
 import awkward as ak
 from awkward._backends.backend import Backend
 from awkward._layout import maybe_posaxis
+from awkward._meta.listmeta import ListMeta
 from awkward._nplikes.array_like import ArrayLike
 from awkward._nplikes.numpy_like import IndexType, NumpyMetadata
 from awkward._nplikes.shape import ShapeItem, unknown_length
@@ -47,7 +48,7 @@ np = NumpyMetadata.instance()
 
 
 @final
-class ListArray(Content):
+class ListArray(ListMeta, Content):
     """
     ListArray generalizes #ak.contents.ListOffsetArray by not
     requiring its `content` to be in increasing order and by allowing it to
@@ -119,7 +120,7 @@ class ListArray(Content):
                     raise AssertionError(where)
     """
 
-    is_list = True
+    _content: Content
 
     def __init__(self, starts, stops, content, *, parameters=None):
         if not isinstance(starts, Index) and starts.dtype in (
