+++ src/awkward/forms/recordform.py
@@ -2,15 +2,14 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable, Iterable, Iterator
+from collections.abc import Callable, Iterable, Iterator, Sequence
 
 import awkward as ak
+from awkward._meta.recordmeta import RecordMeta
 from awkward._nplikes.numpy_like import NumpyMetadata
 from awkward._parameters import type_parameters_equal
-from awkward._regularize import is_integer
-from awkward._typing import DType, JSONSerializable, Self, final
+from awkward._typing import DType, Self, final
 from awkward._util import UNSET
-from awkward.errors import FieldNotFoundError
 from awkward.forms.form import Form, _SpecifierMatcher
 
 __all__ = ("RecordForm",)
@@ -19,8 +18,8 @@ np = NumpyMetadata.instance()
 
 
 @final
-class RecordForm(Form):
-    is_record = True
+class RecordForm(RecordMeta, Form):
+    _contents: Sequence[Form]
 
     def __init__(
         self,
@@ -59,7 +58,7 @@ class RecordForm(Form):
         return self._contents
 
     @property
-    def fields(self):
+    def fields(self) -> list[str]:
         if self._fields is None:
             return [str(i) for i in range(len(self._contents))]
         else:
@@ -91,68 +90,9 @@ class RecordForm(Form):
     ):
         return cls(contents, fields, parameters=parameters, form_key=form_key)
 
-    @property
-    def is_tuple(self):
-        return self._fields is None
-
     def __repr__(self):
         args = [repr(self._contents), repr(self._fields), *self._repr_args()]
         return "{}({})".format(type(self).__name__, ", ".join(args))
-
-    def index_to_field(self, index):
-        if 0 <= index < len(self._contents):
-            if self._fields is None:
-                return str(index)
-            else:
-                return self._fields[index]
-        else:
-            raise IndexError(
-                f"no index {index} in record with {len(self._contents)} fields"
-            )
-
-    def field_to_index(self, field):
-        if self._fields is None:
-            try:
-                i = int(field)
-            except ValueError:
-                pass
-            else:
-                if 0 <= i < len(self._contents):
-                    return i
-        else:
-            try:
-                i = self._fields.index(field)
-            except ValueError:
-                pass
-            else:
-                return i
-        raise FieldNotFoundError(
-            f"no field {field!r} in record with {len(self._contents)} fields"
-        )
-
-    def has_field(self, field):
-        if self._fields is None:
-            try:
-                i = int(field)
-            except ValueError:
-                return False
-            else:
-                return 0 <= i < len(self._contents)
-        else:
-            return field in self._fields
-
-    def content(self, index_or_field):
-        if is_integer(index_or_field):
-            index = index_or_field
-        elif isinstance(index_or_field, str):
-            index = self.field_to_index(index_or_field)
-        else:
-            raise TypeError(
-                "index_or_field must be an integer (index) or string (field), not {}".format(
-                    repr(index_or_field)
-                )
-            )
-        return self._contents[index]
 
     def _to_dict_part(self, verbose, toplevel):
         out = {"class": "RecordArray"}
@@ -194,56 +134,6 @@ class RecordForm(Form):
                 return False
         else:
             return False
-
-    def purelist_parameters(self, *keys: str) -> JSONSerializable:
-        if self._parameters is not None:
-            for key in keys:
-                if key in self._parameters:
-                    return self._parameters[key]
-        return None
-
-    @property
-    def purelist_isregular(self):
-        return True
-
-    @property
-    def purelist_depth(self):
-        return 1
-
-    @property
-    def is_identity_like(self):
-        return False
-
-    @property
-    def minmax_depth(self):
-        if len(self._contents) == 0:
-            return (1, 1)
-        mins, maxs = [], []
-        for content in self._contents:
-            mindepth, maxdepth = content.minmax_depth
-            mins.append(mindepth)
-            maxs.append(maxdepth)
-        return (min(mins), max(maxs))
-
-    @property
-    def branch_depth(self):
-        if len(self._contents) == 0:
-            return (False, 1)
-        anybranch = False
-        mindepth = None
-        for content in self._contents:
-            branch, depth = content.branch_depth
-            if mindepth is None:
-                mindepth = depth
-            if branch or mindepth != depth:
-                anybranch = True
-            if mindepth > depth:
-                mindepth = depth
-        return (anybranch, mindepth)
-
-    @property
-    def dimension_optiontype(self):
-        return False
 
     def _columns(self, path, output, list_indicator):
         for content, field in zip(self._contents, self.fields):
