+++ src/awkward/forms/unionform.py
@@ -2,13 +2,13 @@
 
 from __future__ import annotations
 
-from collections import Counter
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Sequence
 
 import awkward as ak
+from awkward._meta.unionmeta import UnionMeta
 from awkward._nplikes.numpy_like import NumpyMetadata
 from awkward._parameters import type_parameters_equal
-from awkward._typing import DType, Iterator, JSONSerializable, Self, final
+from awkward._typing import DType, Iterator, Self, final
 from awkward._util import UNSET
 from awkward.forms.form import Form, _SpecifierMatcher, index_to_dtype
 
@@ -18,8 +18,8 @@ np = NumpyMetadata.instance()
 
 
 @final
-class UnionForm(Form):
-    is_union = True
+class UnionForm(UnionMeta, Form):
+    _contents: Sequence[Form]
 
     def __init__(
         self,
@@ -154,86 +154,6 @@ class UnionForm(Form):
         ):
             return self._contents == other._contents
 
-        return False
-
-    def purelist_parameters(self, *keys: str) -> JSONSerializable:
-        if self._parameters is not None:
-            for key in keys:
-                if key in self._parameters:
-                    return self._parameters[key]
-
-        for key in keys:
-            out = self._contents[0].purelist_parameter(key)
-            for content in self._contents[1:]:
-                tmp = content.purelist_parameter(key)
-                if out != tmp:
-                    return None
-            return out
-
-        return None
-
-    @property
-    def purelist_isregular(self):
-        for content in self._contents:
-            if not content.purelist_isregular:
-                return False
-        return True
-
-    @property
-    def purelist_depth(self):
-        out = None
-        for content in self._contents:
-            if out is None:
-                out = content.purelist_depth
-            elif out != content.purelist_depth:
-                return -1
-        return out
-
-    @property
-    def is_identity_like(self):
-        return False
-
-    @property
-    def minmax_depth(self):
-        if len(self._contents) == 0:
-            return (0, 0)
-        mins, maxs = [], []
-        for content in self._contents:
-            mindepth, maxdepth = content.minmax_depth
-            mins.append(mindepth)
-            maxs.append(maxdepth)
-        return (min(mins), max(maxs))
-
-    @property
-    def branch_depth(self):
-        if len(self._contents) == 0:
-            return (False, 1)
-        anybranch = False
-        mindepth = None
-        for content in self._contents:
-            branch, depth = content.branch_depth
-            if mindepth is None:
-                mindepth = depth
-            if branch or mindepth != depth:
-                anybranch = True
-            if mindepth > depth:
-                mindepth = depth
-        return (anybranch, mindepth)
-
-    @property
-    def fields(self):
-        field_counts = Counter([f for c in self._contents for f in c.fields])
-        return [f for f, n in field_counts.items() if n == len(self._contents)]
-
-    @property
-    def is_tuple(self):
-        return all(x.is_tuple for x in self._contents) and (len(self._contents) > 0)
-
-    @property
-    def dimension_optiontype(self):
-        for content in self._contents:
-            if content.dimension_optiontype:
-                return True
         return False
 
     def _columns(self, path, output, list_indicator):
